elbepack: elbexml: record source package name and version in source.xml
To analyse a given binary package it is sometimes necessary to also know
the corresponding source package. As the source package can not be
derived from the binary package metadata alone, an explicit link needs
to be established.

Record a binary package's source package name and version in source.xml.
All its other metadata fields can be derived from the binary package.

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>
Reviewed-by: Florian Kauer <florian.kauer@linutronix.de>