elbepack: aptpkgutils: implement recursive package blacklisting in diet builds
Implements package blacklisting for diet builds, this allows the user to
create a list of packages to exclude from the target.
This will recursively avoid installing a package and all its
dependencies on the target even if this package comes as a dependency of
another package.
This must be carefully used as it can create packages with lacking
dependencies.

Signed-off-by: Thomas Bonnefille <thomas.bonnefille@bootlin.com>